+++ CHANGELOG.md
@@ -4,6 +4,7 @@ otj-logging
 2.6.0
 -----
 
+* protect UUID parsing against malformed input
 * added status, url, method getters to RequestLogEvent
 
 2.5.2